src/distance_sensor/requirements.txt
pyzmq==22.0.2
gpiozero==1.6.2